+++ packages/pluggableWidgets/datagrid-text-filter-web/src/components/TextFilterContainer.tsx
@@ -1,11 +1,12 @@
 import { FilterFnList, InputWithFilters } from "@mendix/widget-plugin-filtering/controls";
 import { useBasicSync } from "@mendix/widget-plugin-filtering/helpers/useBasicSync";
-import { useEditableFilterController } from "@mendix/widget-plugin-filtering/helpers/useEditableFilterController";
+import { useStringFilterController } from "@mendix/widget-plugin-filtering/helpers/useStringFilterController";
 import { String_InputFilterInterface } from "@mendix/widget-plugin-filtering/typings/InputFilterInterface";
 import { generateUUID } from "@mendix/widget-plugin-platform/framework/generate-uuid";
 import { observer } from "mobx-react-lite";
 import { createElement, useRef } from "react";
 import { DatagridTextFilterContainerProps, DefaultFilterEnum } from "../../typings/DatagridTextFilterProps";
+import { useOnResetValueEvent, useOnSetValueEvent } from "@mendix/widget-plugin-external-events/hooks";
 
 const filterDefs: Record<DefaultFilterEnum, string> = {
     contains: "Contains",
@@ -29,7 +30,7 @@ const filters: FilterFnList<DefaultFilterEnum> = Object.entries(filterDefs).map(
 );
 
 export interface ContainerProps extends DatagridTextFilterContainerProps {
-    parentChannelName: string | null;
+    parentChannelName: string | undefined;
     filterStore: String_InputFilterInterface;
 }
 
@@ -39,7 +40,7 @@ export const TextFilterContainer: (props: ContainerProps) => React.ReactElement 
 ) {
     const id = (useRef<string>().current ??= `TextFilter${generateUUID()}`);
 
-    const controller = useEditableFilterController({
+    const controller = useStringFilterController({
         filter: props.filterStore,
         defaultFilter: props.defaultFilter,
         defaultValue: props.defaultValue?.value,
@@ -48,6 +49,14 @@ export const TextFilterContainer: (props: ContainerProps) => React.ReactElement 
     });
 
     useBasicSync(props, props.filterStore);
+
+    useOnResetValueEvent({
+        widgetName: props.name,
+        parentChannelName: props.parentChannelName,
+        listener: controller.handleResetValue
+    });
+
+    useOnSetValueEvent({ widgetName: props.name, listener: controller.handleSetValue });
 
     return (
         <InputWithFilters