+++ packages/modules/data-widgets/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+-   We fixed an issue where filters wouldn't reset.
+
 ## [2.24.0] DataWidgets - 2024-09-23
 
 ### [2.10.2] DatagridDateFilter